Debugg everything and create test class to loop through loggable items. Also, create necessary objects in Main.

+++ src/main/java/com/team254/lib/util/ILoggable.java
@@ -1,7 +1,9 @@
 /*what is loggable? --> implimented by subsystems 
  has data to be logged
  implements ILogger.java*/
+import java.util.ArrayList;
+
 public interface ILoggable{
-    ArrayList<float> get_items();
+    ArrayList<Double> get_items();
     ArrayList<String> get_item_names();
 }
